PlaylistsScreen implemented, added CrossRefs to RoomDB

+++ app/src/main/java/com/example/musicwithyou/domain/usecase/playlist_usecase/GetPlaylistUseCase.kt
@@ -7,6 +7,6 @@ class GetPlaylistUseCase @Inject constructor(
     private val repository: PlaylistRepository
 ) {
 
-    suspend operator fun invoke(id: Int) = repository.getPlaylist(id)
+    suspend operator fun invoke(id: Long) = repository.getPlaylist(id)
 
 }
